requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.3
pydantic==2.8.2
python-json-logger==2.0.7
prometheus-client==0.20.0
uvloop==0.19.0
httpx==0.27.0
tenacity==8.3.0
websockets==12.0
binance-connector==3.6.0
